feat: Update unit index table styling
This commit updates the styling of the unit index table in the `index.blade.php` file. Specifically, it modifies the colspan of the last cell in the table to remove the border and background color. This change improves the visual appearance of the table by creating a more seamless transition between the existing rows and any additional content that may be added in the future.