feat: Add "get-kode" route to FormBeliController
This commit adds a new route, "/get-kode", to the FormBeliController. The route is used to retrieve the kode (code) for a specific barang (product) based on the selected merk. This functionality enhances the form-beli feature by allowing users to automatically populate the kode field when selecting a merk for a barang.