feat: Add delete functionality for BarangNama in BarangController
This commit adds the `barang_nama_delete` method to the `BarangController` class. The method allows for the deletion of a `BarangNama` instance, but only if it does not have any associated `Barang` instances. If there are associated `Barang` instances, a redirect back with an error message is returned. This functionality enhances the management of `BarangNama` data by providing a way to delete unused entries.