```text feat: Update select options for "type" field in index.blade.php
This commit updates the select options for the "type" field in the index.blade.php file. The code changes retrieve the select options from the "selectType" variable, which is populated based on the "unitFilter" value. If the "unitFilter" is not empty, the select options are filtered based on the "barang_unit_id" value. Otherwise, all select options are retrieved. This enhancement ensures that the "type" field displays the appropriate options based on the selected unit.