refactor: Improve display of saldo and modal investor in 'index.blade.php'
This commit improves the display of saldo and modal investor in the 'index.blade.php' file. It includes rounding the calculated values for PPN and total, as well as updating the formatting of the displayed values. These changes enhance the accuracy and readability of the information presented to the user.