refactor: Adjust column sizes and add 'SATUAN BARANG' field in 'create.blade.php'
This commit modifies the 'create.blade.php' file in the 'resources/views/db/barang' directory. It adjusts the column sizes for the UNIT, TYPE, KELOMPOK BARANG, NAMA BARANG, KODE BARANG, and MERK BARANG fields, changing them from col-lg-6 col-md-6 to col-lg-3 col-md-3. Additionally, it adds a new field for SATUAN BARANG, allowing the user to select a unit of measurement for the item. These changes improve the layout and usability of the form.