refactor: Add 'invoice_penjualan_detail' route in 'web.php' and corresponding method in 'RekapController.php'
This commit adds a new route and method to handle the 'invoice_penjualan_detail' functionality. In the 'web.php' file, a new route is defined for '/detail/{invoice}' under the 'invoice-penjualan' prefix. This route points to the 'invoice_penjualan_detail' method in the 'RekapController.php' file. The method loads the necessary data for the given invoice and returns the corresponding view. This change enables the user to view the details of a specific invoice in the 'invoice-penjualan' section.